物質の変換仕様を定義
