+++ src/simulations/modular_life_v2/constants.ts
@@ -1,6 +1,7 @@
 import { URLParameterParser } from "../../classes/url_parameter_parser_v3"
 import { isLogLevel, LogLevel } from "./logger"
-import { PhysicalConstant } from "./physics/physical_constant"
+import { defaultMaterialProductionSpec } from "./physics/material"
+import type { PhysicalConstant } from "./physics/physical_constant"
 
 const parameters = new URLParameterParser(document.location.search)
 
@@ -9,6 +10,8 @@ const logLevel: LogLevel = parameters.parseTypedString("log_level", "LogLevel", 
 const physicalConstant: PhysicalConstant = {
   heatLossRate: parameters.parseFloat("heat_loss", { alternativeKey: "ph.h", min: 0 }) ?? 0.25,
   energyHeatConversionRate: parameters.parseFloat("energy_heat_conversion", { alternativeKey: "ph.e", min: 0 }) ?? 0.5,
+
+  materialProduction: defaultMaterialProductionSpec,
 }
 
 export const constants = {
