+++ src/simulations/modular_life_v2/physics/physical_constant.ts
@@ -1,23 +1,8 @@
-export type RedMaterial = "red"
-export type GreenMaterial = "green"
-export type BlueMaterial = "blue"
-export type MaterialType = RedMaterial | GreenMaterial | BlueMaterial
-
-export type Energy = "energy"
-export type Heat = "heat"
-
-export type IngredientType = MaterialType | Energy
-
-type Amount = number
-type ProductionSpec = {
-  /// 1単位の生産に必要な1tick当たりの必要量
-  readonly ingredients: [IngredientType, Amount][]
-  readonly time: number
-}
+import type { MaterialProductionSpec } from "./material"
 
 export type PhysicalConstant = {
   readonly heatLossRate: number
   readonly energyHeatConversionRate: number
 
-  // readonly materialProduction: { [Material in MaterialType]: ProductionSpec }  // TODO:
+  readonly materialProduction: MaterialProductionSpec
 }
